pyproject.toml
[project]
name = "sistema-bancario-fastapi"
version = "0.1.0"
description = ""

authors = [
    {name = "Biz3l",email = "gabrieloliveiralveslegal@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn[latest] (>=0.45.0,<0.46.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "databases (>=0.9.0,<0.10.0)",
    "uuid (>=1.30,<2.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false